test(app): densify naval panel and near-cap test harnesses (#4224)
Slice D: extract capital-link and event-feed scenario tables, consolidate
naval panel part shards (5→3) with shared combine/autoclose harness helpers,
and bring near-cap suites under 600 physical lines.
